Add save button
Red: An assertion failed.
"Assertion Criteria" : ...hasText('')
"Element Matcher" : accessibilityID('messageText')

Now an assertion is failing: we've specified that after clicking save the TextInput should be cleared out, but this isn't happening yet.

In React Web this is the point where we might create a component test to specify this behavior. But Enzyme currently isn't working with React Native 0.51. So we'll go ahead and implement the behavior directly.