Clear text input
We convert the TextInput to a controlled component so we can programmatically set its value to the empty string when Save is tapped. Converting it to a controlled component also results in its value being stored in the state of our custom component, which will be useful for saving that value later.

Red: Cannot find UI Element.
"Element Matcher" : ...hasText('New message')

Now our final assertion fails: we should see the message we entered displayed somewhere on the screen, but we currently don't.